+++ bin/combustor.php
@@ -17,9 +17,13 @@ $app
     ->setCommandPath(__DIR__ . '/../src/Commands')
     ->setCommandNamespace('Rougin\Combustor\Commands');
 
-$app->injector->delegate('Rougin\Describe\Describe', function () {
-    $ci = Rougin\SparkPlug\Instance::create();
+$app->injector->delegate('CI_Controller', function () {
+    return Rougin\SparkPlug\Instance::create();
+});
 
+$ci = $app->injector->make('CI_Controller');
+
+$app->injector->delegate('Rougin\Describe\Describe', function () use ($ci) {
     $ci->load->database();
     $ci->load->helper('inflector');