+++ src/Common/Commands/InstallCommand.php
@@ -6,6 +6,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 use Rougin\Combustor\Common\Tools;
+use Rougin\Combustor\Common\Config;
 use Rougin\Combustor\Commands\AbstractCommand;
 
 /**
@@ -58,26 +59,15 @@ class InstallCommand extends AbstractCommand
      */
     protected function addLibrary($library)
     {
-        $autoload = file_get_contents(APPPATH . 'config/autoload.php');
-        $lines = explode(PHP_EOL, $autoload);
+        $autoload = new Config('autoload', APPPATH . 'config');
 
-        $pattern = '/\$autoload\[\'libraries\'\] = array\((.*?)\)/';
+        $libraries = $autoload->get('libraries', 60, 'array');
 
-        preg_match_all($pattern, $lines[60], $match);
+        if ( ! in_array($library, $libraries)) {
+            array_push($libraries, $library);
 
-        $libraries = explode(', ', end($match[1]));
-
-        if ( ! in_array('\'' . $library . '\'', $libraries)) {
-            array_push($libraries, '\'' . $library . '\'');
-
-            $libraries = array_filter($libraries);
-
-            $pattern = '/\$autoload\[\'libraries\'\] = array\([^)]*\);/';
-            $replacement = '$autoload[\'libraries\'] = array(' . implode(', ', $libraries) . ');';
-
-            $lines[60] = preg_replace($pattern, $replacement, $lines[60]);
-
-            file_put_contents(APPPATH . 'config/autoload.php', implode(PHP_EOL, $lines));
+            $autoload->set('libraries', 60, $libraries, 'array');
+            $autoload->save();
         }
     }
 }