+++ tests/Commands/InstallWildfireCommandTest.php
@@ -4,6 +4,7 @@ namespace Rougin\Combustor\Commands;
 
 use Symfony\Component\Console\Tester\CommandTester;
 
+use Rougin\Combustor\Common\Config;
 use Rougin\Combustor\Fixture\CommandBuilder;
 use Rougin\Combustor\Fixture\CodeIgniterHelper;
 
@@ -12,14 +13,14 @@ use PHPUnit_Framework_TestCase;
 class InstallWildfireCommandTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @var \Symfony\Component\Console\Command\Command
-     */
-    protected $command;
-
-    /**
      * @var string
      */
     protected $appPath;
+
+    /**
+     * @var \Symfony\Component\Console\Command\Command
+     */
+    protected $command;
 
     /**
      * Sets up the command and the application path.
@@ -31,7 +32,6 @@ class InstallWildfireCommandTest extends PHPUnit_Framework_TestCase
         $this->appPath = __DIR__ . '/../TestApp/application';
         $command = 'Rougin\Combustor\Commands\InstallWildfireCommand';
 
-        CodeIgniterHelper::setDefaults($this->appPath);
         $this->command = CommandBuilder::create($command);
     }
 
@@ -42,12 +42,31 @@ class InstallWildfireCommandTest extends PHPUnit_Framework_TestCase
      */
     public function testFileExists()
     {
+        CodeIgniterHelper::setDefaults($this->appPath);
+
         $command = new CommandTester($this->command);
         $command->execute([]);
 
         $file = $this->appPath . '/libraries/Wildfire.php';
 
         $this->assertFileExists($file);
+
+        $autoload = new Config('autoload', $this->appPath . '/config');
+
+        $drivers = [ 'session' ];
+        $helpers = [ 'form', 'url' ];
+
+        $this->assertEquals($drivers, $autoload->get('drivers', 75, 'array'));
+        $this->assertEquals($helpers, $autoload->get('helper', 85, 'array'));
+
+        $config = new Config('config', $this->appPath . '/config');
+
+        $composerAutoload = 'realpath(\'vendor\') . \'/autoload.php\'';
+
+        $this->assertEquals($composerAutoload, $config->get('composer_autoload', 132, 'string'));
+        $this->assertEmpty($config->get('index_page', 31, 'string'));
+        $this->assertEquals(md5('rougin'), $config->get('encryption_key', 310, 'string'));
+
         CodeIgniterHelper::setDefaults($this->appPath);
     }
 }