+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-All Notable changes to `Combustor` will be documented in this file
+All Notable changes to `Combustor` will be documented in this file.
 
 ## [Unreleased](https://github.com/rougin/combustor/compare/v1.1.4...HEAD)
 
 ### Added
 - Tests
+- Contributor Code of Conduct
 
 ### Changed
 - Commands to be testable